requirements.txt
fastapi[standard]
heyoo
llama_index
python-dotenv
fpdf2
pandas
matplotlib
llama-index-llms-google-genai